requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.6.3
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
Faker==0.9.1
future==0.16.0
idna==2.7
pycparser==2.19
python-dateutil==2.7.3
python-dotenv==0.9.1
python-telegram-bot==11.1.0
requests==2.19.1
six==1.11.0
text-unidecode==1.2
urllib3==1.23
